+++ src/main/java/com/tristankechlo/toolleveling/client/TooltipEvents.java
@@ -1,0 +1,26 @@
+package com.tristankechlo.toolleveling.client;
+
+import com.tristankechlo.toolleveling.utils.Utils;
+
+import net.minecraft.client.gui.screen.Screen;
+import net.minecraft.item.Item;
+import net.minecraft.util.text.StringTextComponent;
+import net.minecraft.util.text.TextFormatting;
+import net.minecraft.util.text.TranslationTextComponent;
+import net.minecraftforge.event.entity.player.ItemTooltipEvent;
+import net.minecraftforge.eventbus.api.SubscribeEvent;
+
+public final class TooltipEvents {
+
+	@SubscribeEvent
+	public void onTooltip(final ItemTooltipEvent event) {
+		if (event.getFlags().isAdvanced() && Screen.hasShiftDown()) {
+			final Item item = event.getItemStack().getItem();
+			final long worth = Utils.getItemWorth(item);
+			event.getToolTip().add(new StringTextComponent(""));
+			event.getToolTip().add(new TranslationTextComponent("tooltip.toolleveling.itemworth", worth)
+					.withStyle(TextFormatting.AQUA));
+		}
+	}
+
+}